server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "uvicorn",
    "requests",
    "python-dotenv",
    "fastapi",
    "mcp[cli]>=1.12.3",
]

[project.scripts]
run = "uvicorn server.server:mcp --reload"
